Refactor route handlers
